+++ constants/Money.ts
@@ -2,4 +2,5 @@ import Intl from 'intl';
 import 'intl/locale-data/jsonp/en';
 const moneyFormatter = new Intl.NumberFormat('vi-VN');
 
-export const moneyFormat = (money: number): string => `${moneyFormatter.format(money)}đ`;
+export const moneyFormat = (money: number, includeMoneyCharacter = true): string =>
+  `${moneyFormatter.format(money)}${includeMoneyCharacter ? 'đ' : ''}`;