+++ constants/interface.ts
@@ -43,6 +43,8 @@ export interface IBook {
   description: string;
   images: string[];
   categoryIds: number[];
+  numOfCopies: number;
+  barcode?: string;
 }
 
 export interface IBookResponse extends BaseEntity {
@@ -64,6 +66,8 @@ export interface IOrder {
   store: IStore;
   orderDetails: Array<IResponseOrderDetail>;
   totalAmount: number;
+  status: OrderStatus;
+  user: IUser;
 }
 
 export interface IResponseOrderDetail {
@@ -79,21 +83,20 @@ export enum OrderDetailAction {
   REMOVE = 'remove',
 }
 
-export interface IBookCopy {
-  barcode: string;
-  status: BookStatus;
-}
-
 export interface IBookLoanResponse extends BaseEntity {
   id: number;
   dueDate: string;
-  order: Omit<IOrder, 'orderDetails' | 'store'>;
-  bookCopy: IBookCopy;
+  order: IOrder;
+  book: IBook;
+}
+
+export interface IOrderBookLoanResponse {
+  order: IOrder;
+  bookLoans: IBookLoanResponse[];
 }
 
 export interface IReturnBook {
   status: ReturnBookType;
-  barcode: string;
 }
 export interface IOrderDetail {
   quantity: number;