+++ screens/Account/AccountScreen.tsx
@@ -5,14 +5,16 @@ import { TouchableOpacity } from 'react-native';
 import { useNavigation } from '@react-navigation/native';
 import { mainColor } from '../../constants/Colors';
 import { useAppDispatch, useAppSelector } from '../../app/hook';
-import { logoutAction, selectOwnStore, selectUser } from '../../reducers/authSlice';
+import { logoutAction, profileAction, selectOwnStore, selectUser } from '../../reducers/authSlice';
 import { RootStackScreenProps } from '../../types';
+import { moneyFormat } from '../../constants/Money';
 
 const AccountScreen: React.FC<RootStackScreenProps<'Root'>> = ({ navigation }) => {
   const nav = useNavigation<any>();
   const dispatch = useAppDispatch();
   const user = useAppSelector(selectUser);
   const ownStore = useAppSelector(selectOwnStore);
+  const profile = useAppSelector(selectUser);
 
   const handleLogout = () => {
     Alert.alert('Đăng xuất', 'Bạn có chắc muốn đăng xuất?', [
@@ -95,7 +97,7 @@ const AccountScreen: React.FC<RootStackScreenProps<'Root'>> = ({ navigation }) =
             <TouchableOpacity style={styles.item} activeOpacity={0.8} onPress={() => navigation.navigate('Renting')}>
               <View
                 style={{
-                  backgroundColor: '#000000',
+                  backgroundColor: '#ff9b30',
                   borderRadius: 5,
                   padding: 5,
                   width: 30,
@@ -135,6 +137,29 @@ const AccountScreen: React.FC<RootStackScreenProps<'Root'>> = ({ navigation }) =
                 <Text style={styles.text}>Yêu thích</Text>
               </View>
               <Icon name="angle-right" type="font-awesome" style={{ flex: 1 }} color="#c1c1c1" />
+            </TouchableOpacity>
+
+            <TouchableOpacity style={styles.item} activeOpacity={0.8} onPress={() => dispatch(profileAction())}>
+              <View
+                style={{
+                  backgroundColor: '#ff00e885',
+                  borderRadius: 5,
+                  padding: 5,
+                  width: 30,
+                  height: 30,
+                }}
+              >
+                <Icon name="money" type="font-awesome" color="#fff" size={20} solid={true} />
+              </View>
+              <View
+                style={{
+                  flex: 2,
+                  marginLeft: 30,
+                }}
+              >
+                <Text style={styles.text}>WibuCoin: {moneyFormat(profile?.coin || 0, false)}</Text>
+              </View>
+              {/* <Icon name="money" type="font-awesome" style={{ flex: 1 }} color="#c1c1c1" /> */}
             </TouchableOpacity>
             <TouchableOpacity style={styles.item} activeOpacity={0.8}>
               <View