+++ screens/Home/HomeScreen.tsx
@@ -14,7 +14,6 @@ export default function HomeScreen() {
   const navigation = useNavigation();
   const dispatch = useAppDispatch();
   const categories = useAppSelector(selectData);
-  console.log(categories);
 
   useEffect(() => {
     dispatch(GetAllCategoryAction());